flask-backend/requirements.txt
Flask
Flask-Cors
textblob
python-dotenv
